# adapters/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.LoginUserInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.LoginUserOutputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/batch": {
            "post": {
                "description": "Create batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Create batch",
                "parameters": [
                    {
                        "description": "batch",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/batch.CreateBatchInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/batch.CreateBatchInputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/brewery": {
            "get": {
                "description": "Get brewery by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brewery"
                ],
                "summary": "Get brewery by user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "brewery",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/brewery.GetUserBreweriesOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create brewery",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brewery"
                ],
                "summary": "Create brewery",
                "parameters": [
                    {
                        "description": "brewery",
                        "name": "brewery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/brewery.CreateBreweryInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/brewery.CreateBreweryInputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/brewery/invite": {
            "post": {
                "description": "Invite user to a brewery",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brewery"
                ],
                "summary": "Invite user",
                "parameters": [
                    {
                        "description": "invite",
                        "name": "invite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/brewery.InviteUserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/brewery.InviteUserOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipes": {
            "post": {
                "description": "Create recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Create recipe",
                "parameters": [
                    {
                        "description": "recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recipe.CreateRecipeInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/recipe.CreateRecipeInputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/start-batch": {
            "post": {
                "description": "Create batch step",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Create batch step",
                "parameters": [
                    {
                        "description": "batch step",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/batch.CreateBatchStepInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/batch.CreateBatchStepInputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserInputDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HttpErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "batch.CreateBatchInputDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "recipe_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "batch.CreateBatchStepInputDto": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "recipe_step_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "brewery.CreateBreweryInputDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "brewery.GetUserBreweriesOutputDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "brewery.InviteUserInputDTO": {
            "type": "object",
            "properties": {
                "brewery_id": {
                    "type": "string"
                },
                "invited_user_id": {
                    "type": "string"
                },
                "inviting_user_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "brewery.InviteUserOutputDTO": {
            "type": "object",
            "properties": {
                "brewery_id": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "invited_user_id": {
                    "type": "string"
                },
                "inviting_user_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "common.HttpErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "recipe.CreateRecipeInputDto": {
            "type": "object",
            "properties": {
                "brewery_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/recipe.Steps"
                    }
                }
            }
        },
        "recipe.Steps": {
            "type": "object",
            "properties": {
                "instruction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "user.CreateUserInputDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.LoginUserInputDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.LoginUserOutputDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}